Module Fmlib_std

Standard Data Types.

Common Interfaces

module Interfaces : sig ... end

Standard Module Types

Data Types

module Array : sig ... end

A thin wrapper around Stdlib.Array with additional functions and sets and maps based on arrays

module Btree : sig ... end

Sets and maps based on B trees.

module Deque : sig ... end

A double ended queue. Insert from both ends, pop only from the front.

module Rb_array : sig ... end

A Radix Balanced Array. A functional long array which allows fast random access, fast insertion and deletion at the rear end.

module Int : sig ... end

Module to represent the type int

module List : sig ... end

A thin wrapper around Stdlib.List which avoids throwing exceptions and with some additional monadic functions.

module Option : sig ... end

Optional elements of a certain type.

module Result : sig ... end

Result: Handling results of operations which can fail

module String : sig ... end

A small wrapper around Stdlib.String with some extra functions.

module Void : sig ... end

Void.t is the type of an object which cannot exist.