Module Map.Source

Module which satisfies the interface Interfaces.SOURCE

Parameters

module Value : Interfaces.ANY

Signature

type 'a map = 'a t

Standard API

include Interfaces.SOURCE with type item = Key.t * Value.t

type item = Key.t * Value.t

Type of items coming from the source.

type t

t A stream of items.

val has_more : t -> bool

has_more s Does the stream s have more items?

val peek : t -> item

peek s Peek the next item from the stream s.

Precondition: has_more s

val advance : t -> t

advance s Pop the top item off the stream s.

Precondition: has_more s

Create a stream

val make : Value.t map -> t