Module Fmlib_std.Interfaces

Standard Module Types

module type ANY = sig ... end

Interface for a module which has a type t.

module type MAP = sig ... end

Interface for a finite map i.e. a set of key value pairs.

module type MONAD = sig ... end

Interface for a module with a monadic container.

module type SET = sig ... end

Interface for a finite set.

module type SORTABLE = sig ... end

Interface for a sortable type. A type is sortable, if it has a comparison function.

module type SOURCE = sig ... end

Interface for a source of items i.e. a stream.

module type SINK = sig ... end

Interface for a sink of items.