Module Fmlib_std.Unicode

Support for unicode

Utf8 Encoding

Unicode has 17 planes with `2^16` code points minus `2^11` technically invalid surrogates.

Ascii characters i.e. values in the range `0..127` or `0b0000_0000..0b0xxx_xxxx` are valid utf8 characters.

Code points: `U+0000`- `U+10FFFF` . These are `0x11_0000` (`2^20 + 2^16 = 1_048_576 + 65_536 = 1_114_112`) code points. However the code points `U+D800`-`U+DFFF` (`2^11 = 2048`) are used to code *surrogate pairs* and are therefore invalid. I.e. there are only `1_112_064` valid code points.

UTF8 code points: `0x0000-0xD7FF`, `0xE000-0x10FFFF`.

Replacement character `U+FFFD` for uninterpretable encoding.

Encodings:

        +--------+-------------------------------------+-------------------+
        | 1 byte | 0xxxxxxx                            | 7 bits            |
        | 2 byte | 110xxxxx 10xxxxxx                   | 11 bits (5 + 6)   |
        | 3 byte | 1110xxxx 10xxxxxx 10xxxxxx          | 16 bits (4 + 2*6) |
        | 4 byte | 11110xxx 10xxxxxx 10xxxxxx 10xxxxxx | 21 bits (3 + 3*6) |
        +--------+-------------------------------------+-------------------+

There are Advantages:

Backward compatible to ascii.

Easy recognition of multibyte boundaries. Leading bytes and continuation bytes do not share values.

First byte in a multibyte sequence indicates the number of bytes of the multibyte sequence (number of binary `1`s in the first byte.

val is_utf8_1byte : char -> bool

is_utf8_1byte c Is the character c a one byte character in utf8?

val is_utf8_2byte : char -> bool

is_utf8_2byte c Does the character c start a utf8 2 byte character?

val is_utf8_3byte : char -> bool

is_utf8_3byte c Does the character c start a utf8 3 byte character?

val is_utf8_4byte : char -> bool

is_utf8_4byte c Does the character c start a utf8 4 byte character?

val is_utf8_continuation : char -> bool

is_utf8_4byte c Is the character c start a utf8 continuation character?

module Utf8 : sig ... end

Support for scanning bytes of a character in utf8 encoding.