Module Unicode.Utf8

Support for scanning bytes of a character in utf8 encoding.

type t

Partially scanned utf8 character.

val is_complete : t -> bool

is_complete u Does u contain a complete (valid or invalid) character in utf8 encoding?

val uchar : t -> Stdlib.Uchar.t

Return the unicode character of a completely scanned utf8 encoded character. If there is not valid unicode character then return Uchar.rep.

val start : t

The start of a scan of a utf8 encoded character.

val put_char : char -> t -> t

Feed the scanner with the next byte of a ut8 encoded character stream.