Module String.To_source

Conversion of a string to a source of characters.

include Interfaces.SOURCE with type item = char

type item = char

Type of items coming from the source.

type t

t A stream of items.

val has_more : t -> bool

has_more s Does the stream s have more items?

val peek : t -> item

peek s Peek the next item from the stream s.

Precondition: has_more s

val advance : t -> t

advance s Pop the top item off the stream s.

Precondition: has_more s

val make : string -> t

make str A character stream generated from the string str.